@@ -6,8 +6,12 @@
     "homepage": "http://developer.constantcontact.com",
     "license": "MIT",
     "require": {
-        "php": ">=5.3.0",
-        "ext-curl": "*"
+        "php": ">=5.4.0",
+        "ext-curl": "*",
+        "guzzlehttp/guzzle": "5.1.0"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "4.4.1"
     },
     "authors": [
         {
