@@ -1,61 +1,49 @@
 <?php
 
-use Ctct\Services\EmailMarketingService;
-use Ctct\Util\RestClient;
-use Ctct\Util\CurlResponse;
+use Ctct\Components\ResultSet;
 use Ctct\Components\EmailMarketing\Campaign;
+
+use GuzzleHttp\Client;
+use GuzzleHttp\Exception\ClientException;
+use GuzzleHttp\Subscriber\Mock;
+use GuzzleHttp\Stream\Stream;
+use GuzzleHttp\Message\Response;
 
 class EmailMarketingServiceUnitTest extends PHPUnit_Framework_TestCase
 {
-    private $restClient;
-    private $emailMarketingService;
+    /**
+     * @var Client
+     */
+    private static $client;
 
-    public function setUp()
+    public static function setUpBeforeClass()
     {
-        $this->restClient = $this->getMock('Ctct\Util\RestClientInterface');
-        $this->emailMarketingService = new EmailMarketingService("apikey", $this->restClient);
-    }
-
-    public function testGetCampaignsModifiedSince()
-    {
-        $curlResponse = CurlResponse::create(JsonLoader::getCampaignModifiedSinceJson(1), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
-
-        $response = $this->emailMarketingService->getCampaigns(
-            'access_token',
-            array('modified_since' => '2013-01-12T20:04:59.436Z', 'limit' => 2)
-        );
-        $campaigns = $response->results;
-
-        $this->assertInstanceOf('Ctct\Components\EmailMarketing\Campaign', $campaigns[0]);
-        $this->assertEquals("ABcGFnZU51bT0yJnBhZ2VTaXplPTImbW9kaWZpZWRfc2luY2U9MTM1OTUxNjYzMDU5MA", $response->next);
-        $this->assertEquals("9112921497760", $campaigns[0]->id);
-        $this->assertEquals("Email Created 2013/03/29, 11:30 PM", $campaigns[0]->name);
-        $this->assertEquals("DRAFT", $campaigns[0]->status);
-        $this->assertEquals("2013-03-30T03:30:48.033Z", $campaigns[0]->modified_date);
-
-        $this->assertEquals("9112756952331", $campaigns[1]->id);
-        $this->assertEquals("CampaignName234", $campaigns[1]->name);
-        $this->assertEquals("DRAFT", $campaigns[1]->status);
-        $this->assertEquals("2013-03-14T15:00:07.883Z", $campaigns[1]->modified_date);
+        self::$client = new Client();
+        $getCampaignsStream = Stream::factory(JsonLoader::getCampaignsJson());
+        $getCampaignStream = Stream::factory(JsonLoader::getCampaignJson());
+        $mock = new Mock([
+            new Response(200, array(), $getCampaignsStream),
+            new Response(204, array()),
+            new Response(400, array()),
+            new Response(200, array(), $getCampaignStream),
+            new Response(201, array(), $getCampaignStream),
+            new Response(200, array(), $getCampaignStream)
+        ]);
+        self::$client->getEmitter()->attach($mock);
     }
 
     public function testGetCampaigns()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getCampaignsJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/')->json();
+        $result = new ResultSet($response['results'], $response['meta']);
+        $campaigns = array();
+        foreach ($result->results as $campaign) {
+            $campaigns[] = Campaign::create($campaign);
+        }
 
-        $response = $this->emailMarketingService->getCampaigns('access_token');
-        $campaigns = $response->results;
-
+        $this->assertInstanceOf('Ctct\Components\ResultSet', $result);
         $this->assertInstanceOf('Ctct\Components\EmailMarketing\Campaign', $campaigns[0]);
-        $this->assertEquals("cGFnZU51bT0yJnBhZ2VTaXplPTM", $response->next);
+        $this->assertEquals("cGFnZU51bT0yJnBhZ2VTaXplPTM", $result->next);
         $this->assertEquals("1100371240640", $campaigns[0]->id);
         $this->assertEquals("Email Created 2012/11/29, 4:13 PM", $campaigns[0]->name);
         $this->assertEquals("SENT", $campaigns[0]->status);
@@ -69,38 +57,26 @@ class EmailMarketingServiceUnitTest extends PHPUnit_Framework_TestCase
 
     public function testDeleteCampaign()
     {
-        $curlResponse = CurlResponse::create(null, array('http_code' => 204));
-        $this->restClient->expects($this->once())
-            ->method('delete')
-            ->with()
-            ->will($this->returnValue($curlResponse));
-
-        $response = $this->emailMarketingService->deleteCampaign('access_token', "1100368835463");
-        $this->assertTrue($response);
+        $response = self::$client->delete('/');
+        $this->assertEquals(204, $response->getStatusCode());
     }
 
     public function testDeleteCampaignFailed()
     {
-        $curlResponse = CurlResponse::create(null, array('http_code' => 400));
-        $this->restClient->expects($this->once())
-            ->method('delete')
-            ->with()
-            ->will($this->returnValue($curlResponse));
-
-        $response = $this->emailMarketingService->deleteCampaign('access_token', "1100368835463");
-        $this->assertEquals(false, $response);
+        try {
+            self::$client->delete('/');
+            $this->fail("Delete did not fail");
+        } catch (ClientException $e) {
+            $this->assertEquals(400, $e->getCode());
+        }
     }
 
     public function testGetCampaign()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getCampaignJson(), array('http_code' => 201));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $campaign = $this->emailMarketingService->getCampaign('access_token', 11109369315398);
-
+        $campaign = Campaign::create($response->json());
+        $this->assertInstanceOf('Ctct\Components\EmailMarketing\Campaign', $campaign);
         $this->assertEquals("1100394165290", $campaign->id);
         $this->assertEquals("CampaignName-05965ddb-12d2-43e5-b8f3-0c22ca487c3a", $campaign->name);
         $this->assertEquals("CampaignSubject", $campaign->subject);
@@ -165,14 +141,10 @@ class EmailMarketingServiceUnitTest extends PHPUnit_Framework_TestCase
 
     public function testAddCampaign()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getCampaignJson(), array('http_code' => 201));
-        $this->restClient->expects($this->once())
-            ->method('post')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->post('/');
 
-        $campaign = $this->emailMarketingService->addCampaign('access_token', new Campaign());
-
+        $campaign = Campaign::create($response->json());
+        $this->assertInstanceOf('Ctct\Components\EmailMarketing\Campaign', $campaign);
         $this->assertEquals("1100394165290", $campaign->id);
         $this->assertEquals("CampaignName-05965ddb-12d2-43e5-b8f3-0c22ca487c3a", $campaign->name);
         $this->assertEquals("CampaignSubject", $campaign->subject);
@@ -236,14 +208,10 @@ class EmailMarketingServiceUnitTest extends PHPUnit_Framework_TestCase
 
     public function testUpdateCampaign()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getCampaignJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('put')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->put('/');
 
-        $campaign = $this->emailMarketingService->updateCampaign('access_token', new Campaign());
-
+        $campaign = Campaign::create($response->json());
+        $this->assertInstanceOf('Ctct\Components\EmailMarketing\Campaign', $campaign);
         $this->assertEquals("1100394165290", $campaign->id);
         $this->assertEquals("CampaignName-05965ddb-12d2-43e5-b8f3-0c22ca487c3a", $campaign->name);
         $this->assertEquals("CampaignSubject", $campaign->subject);
