@@ -1,75 +1,62 @@
 <?php
 
-use Ctct\Util\CurlResponse;
-use Ctct\Services\ListService;
-use Ctct\Util\RestClient;
+use Ctct\Components\ResultSet;
+use Ctct\Components\Contacts\Contact;
 use Ctct\Components\Contacts\ContactList;
+
+use GuzzleHttp\Client;
+use GuzzleHttp\Subscriber\Mock;
+use GuzzleHttp\Stream\Stream;
+use GuzzleHttp\Message\Response;
 
 class ListServiceUnitTest extends PHPUnit_Framework_TestCase
 {
-    private $restClient;
-    private $listService;
+    /**
+     * @var Client
+     */
+    private static $client;
 
-    public function setUp()
+    public static function setUpBeforeClass()
     {
-        $this->restClient = $this->getMock('Ctct\Util\RestClientInterface');
-        $this->listService = new ListService("apikey", $this->restClient);
+        self::$client = new Client();
+        $getListsStream = Stream::factory(JsonLoader::getListsJson());
+        $getListStream = Stream::factory(JsonLoader::getListJson());
+        $getContactsStream = Stream::factory(JsonLoader::getContactsJson());
+        $mock = new Mock([
+            new Response(200, array(), $getListsStream),
+            new Response(200, array(), $getListStream),
+            new Response(201, array(), $getListStream),
+            new Response(200, array(), $getListStream),
+            new Response(200, array(), $getContactsStream)
+        ]);
+        self::$client->getEmitter()->attach($mock);
     }
 
     public function testGetLists()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getListsJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $response = $this->listService->getLists('access_token');
+        $lists = array();
+        foreach ($response->json() as $list) {
+            $lists[] = ContactList::create($list);
+        }
+        $this->assertInstanceOf('Ctct\Components\Contacts\ContactList', $lists[0]);
+        $this->assertEquals(1, $lists[0]->id);
+        $this->assertEquals("General Interest", $lists[0]->name);
+        $this->assertEquals("ACTIVE", $lists[0]->status);
+        $this->assertEquals(17, $lists[0]->contact_count);
 
-        $this->assertInstanceOf("Ctct\Components\Contacts\ContactList", $response[0]);
-        $this->assertEquals(1, $response[0]->id);
-        $this->assertEquals("General Interest", $response[0]->name);
-        $this->assertEquals("ACTIVE", $response[0]->status);
-        $this->assertEquals(17, $response[0]->contact_count);
-
-        $this->assertEquals(3, $response[1]->id);
-        $this->assertEquals("mod_Test List 1", $response[1]->name);
-        $this->assertEquals("HIDDEN", $response[1]->status);
-        $this->assertEquals(18, $response[1]->contact_count);
-    }
-
-    public function testGetListsModifiedSince()
-    {
-        $curlResponse = CurlResponse::create(JsonLoader::getListsJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
-
-        $response = $this->listService->getLists('access_token', array('modified_since' => '2013-01-12T20:04:59.436Z'));
-
-        $this->assertInstanceOf("Ctct\Components\Contacts\ContactList", $response[0]);
-        $this->assertEquals(1, $response[0]->id);
-        $this->assertEquals("General Interest", $response[0]->name);
-        $this->assertEquals("ACTIVE", $response[0]->status);
-        $this->assertEquals(17, $response[0]->contact_count);
-
-        $this->assertEquals(3, $response[1]->id);
-        $this->assertEquals("mod_Test List 1", $response[1]->name);
-        $this->assertEquals("HIDDEN", $response[1]->status);
-        $this->assertEquals(18, $response[1]->contact_count);
+        $this->assertEquals(3, $lists[1]->id);
+        $this->assertEquals("mod_Test List 1", $lists[1]->name);
+        $this->assertEquals("HIDDEN", $lists[1]->status);
+        $this->assertEquals(18, $lists[1]->contact_count);
     }
 
     public function testGetList()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getListJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
-
-        $list = $this->listService->getList('access_token', 6);
-        $this->assertInstanceOf("Ctct\Components\Contacts\ContactList", $list);
+        $response = self::$client->get('/');
+        $list = ContactList::create($response->json());
+        $this->assertInstanceOf('Ctct\Components\Contacts\ContactList', $list);
         $this->assertEquals(6, $list->id);
         $this->assertEquals("Test List 4", $list->name);
         $this->assertEquals("HIDDEN", $list->status);
@@ -78,14 +65,9 @@ class ListServiceUnitTest extends PHPUnit_Framework_TestCase
 
     public function testAddList()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getListJson(), array('http_code' => 204));
-        $this->restClient->expects($this->once())
-            ->method('post')
-            ->with()
-            ->will($this->returnValue($curlResponse));
-
-        $list = $this->listService->addList('access_token', new ContactList());
-        $this->assertInstanceOf("Ctct\Components\Contacts\ContactList", $list);
+        $response = self::$client->post('/');
+        $list = ContactList::create($response->json());
+        $this->assertInstanceOf('Ctct\Components\Contacts\ContactList', $list);
         $this->assertEquals(6, $list->id);
         $this->assertEquals("Test List 4", $list->name);
         $this->assertEquals("HIDDEN", $list->status);
@@ -94,14 +76,9 @@ class ListServiceUnitTest extends PHPUnit_Framework_TestCase
 
     public function testUpdateList()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getListJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('put')
-            ->with()
-            ->will($this->returnValue($curlResponse));
-
-        $list = $this->listService->updateList('access_token', new ContactList());
-        $this->assertInstanceOf("Ctct\Components\Contacts\ContactList", $list);
+        $response = self::$client->put('/');
+        $list = ContactList::create($response->json());
+        $this->assertInstanceOf('Ctct\Components\Contacts\ContactList', $list);
         $this->assertEquals(6, $list->id);
         $this->assertEquals("Test List 4", $list->name);
         $this->assertEquals("HIDDEN", $list->status);
@@ -110,23 +87,18 @@ class ListServiceUnitTest extends PHPUnit_Framework_TestCase
 
     public function testGetContactsFromList()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getContactsJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/')->json();
+        $result = new ResultSet($response['results'], $response['meta']);
+        $this->assertInstanceOf('Ctct\Components\ResultSet', $result);
 
-        $response = $this->listService->getContactsFromList('access_token', 1);
-        $this->assertInstanceOf("Ctct\Components\ResultSet", $response);
+        $contact = Contact::create($result->results[1]);
 
-        $contact = $response->results[1];
-
+        $this->assertInstanceOf('Ctct\Components\Contacts\Contact', $contact);
         $this->assertEquals(231, $contact->id);
         $this->assertEquals("ACTIVE", $contact->status);
         $this->assertEquals("", $contact->fax);
         $this->assertEquals("", $contact->prefix_name);
         $this->assertEquals("Jimmy", $contact->first_name);
-        $this->assertEquals("", $contact->middle_name);
         $this->assertEquals("Roving", $contact->last_name);
         $this->assertEquals("Bear Tamer", $contact->job_title);
         $this->assertEquals("Animal Trainer Pro", $contact->company_name);
@@ -135,10 +107,12 @@ class ListServiceUnitTest extends PHPUnit_Framework_TestCase
         $this->assertEquals("", $contact->source);
 
         // custom fields
+        $this->assertInstanceOf('Ctct\Components\Contacts\CustomField', $contact->custom_fields[0]);
         $this->assertEquals("CustomField1", $contact->custom_fields[0]->name);
         $this->assertEquals("1", $contact->custom_fields[0]->value);
 
         //addresses
+        $this->assertInstanceOf('Ctct\Components\Contacts\Address', $contact->addresses[0]);
         $this->assertEquals("Suite 101", $contact->addresses[0]->line1);
         $this->assertEquals("line2", $contact->addresses[0]->line2);
         $this->assertEquals("line3", $contact->addresses[0]->line3);
@@ -153,10 +127,12 @@ class ListServiceUnitTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(0, count($contact->notes));
 
         //lists
+        $this->assertInstanceOf('Ctct\Components\Contacts\ContactList', $contact->lists[0]);
         $this->assertEquals(1, $contact->lists[0]->id);
         $this->assertEquals("ACTIVE", $contact->lists[0]->status);
 
         // EmailAddress
+        $this->assertInstanceOf('Ctct\Components\Contacts\EmailAddress', $contact->email_addresses[0]);
         $this->assertEquals("ACTIVE", $contact->email_addresses[0]->status);
         $this->assertEquals("NO_CONFIRMATION_REQUIRED", $contact->email_addresses[0]->confirm_status);
         $this->assertEquals("ACTION_BY_OWNER", $contact->email_addresses[0]->opt_in_source);
