@@ -14,13 +14,12 @@ http://constantcontact.mashery.com.
 -->
 
 <?php
-// require the autoloader
+// require the autoloaders
 require_once '../src/Ctct/autoload.php';
+require_once '../vendor/autoload.php';
 
 use Ctct\ConstantContact;
 use Ctct\Components\Contacts\Contact;
-use Ctct\Components\Contacts\ContactList;
-use Ctct\Components\Contacts\EmailAddress;
 use Ctct\Exceptions\CtctException;
 
 // Enter your Constant Contact APIKEY and ACCESS_TOKEN
@@ -31,10 +30,13 @@ $cc = new ConstantContact(APIKEY);
 
 // attempt to fetch lists in the account, catching any exceptions and printing the errors to screen
 try {
-    $lists = $cc->getLists(ACCESS_TOKEN);
+    $lists = $cc->listService->getLists(ACCESS_TOKEN);
 } catch (CtctException $ex) {
     foreach ($ex->getErrors() as $error) {
         print_r($error);
+    }
+    if (!isset($lists)) {
+        $lists = null;
     }
 }
 
@@ -42,8 +44,8 @@ try {
 if (isset($_POST['email']) && strlen($_POST['email']) > 1) {
     $action = "Getting Contact By Email Address";
     try {
-        // check to see if a contact with the email addess already exists in the account
-        $response = $cc->getContactByEmail(ACCESS_TOKEN, $_POST['email']);
+        // check to see if a contact with the email address already exists in the account
+        $response = $cc->contactService->getContacts(ACCESS_TOKEN, array("email" => $_POST['email']));
 
         // create a new contact if one does not exist
         if (empty($response->results)) {
@@ -62,13 +64,13 @@ if (isset($_POST['email']) && strlen($_POST['email']) > 1) {
              *
              * See: http://developer.constantcontact.com/docs/contacts-api/contacts-index.html#opt_in
              */
-            $returnContact = $cc->addContact(ACCESS_TOKEN, $contact, true);
+            $returnContact = $cc->contactService->addContact(ACCESS_TOKEN, $contact);
 
             // update the existing contact if address already existed
         } else {
             $action = "Updating Contact";
 
-            $contact = $response->results[0];
+            $contact = Contact::create($response->results[0]);
             $contact->addList($_POST['list']);
             $contact->first_name = $_POST['first_name'];
             $contact->last_name = $_POST['last_name'];
@@ -80,7 +82,7 @@ if (isset($_POST['email']) && strlen($_POST['email']) > 1) {
              *
              * See: http://developer.constantcontact.com/docs/contacts-api/contacts-index.html#opt_in
              */
-            $returnContact = $cc->updateContact(ACCESS_TOKEN, $contact, true);
+            $returnContact = $cc->contactService->updateContact(ACCESS_TOKEN, $contact);
         }
 
         // catch any exceptions thrown during the process and print the errors to screen
