@@ -1,14 +1,12 @@
 <?php
 namespace Ctct\Exceptions;
 
-use Exception;
-
 /**
  * Exception thrown if an error occurred during OAuth2 authentication process
  *
  * @package     exceptions
  * @author         djellesma
  */
-class OAuth2Exception extends Exception
+class OAuth2Exception extends CtctException
 {
 }
