@@ -1,8 +1,10 @@
 <?php
 namespace Ctct\Services;
 
-use Ctct\Util\RestClient;
-use Ctct\Util\RestClientInterface;
+use Ctct\Exceptions\CtctException;
+use Ctct\Util\Config;
+use GuzzleHttp\Client;
+use GuzzleHttp\Exception\ClientException;
 
 /**
  * Super class for all services
@@ -13,78 +15,68 @@ use Ctct\Util\RestClientInterface;
 abstract class BaseService
 {
     /**
-     * RestClient Implementation to use for HTTP requests
-     * @var RestClientInterface
+     * Helper function to return required headers for making an http request with constant contact
+     * @param $accessToken - OAuth2 access token to be placed into the Authorization header
+     * @return array - authorization headers
      */
-    protected $restClient;
+    private static function getHeaders($accessToken)
+    {
+        return array(
+            'User-Agent' => 'ConstantContact AppConnect PHP Library v' . Config::get('settings.version'),
+            'Content-Type' => 'application/json',
+            'Accept' => 'application/json',
+            'Authorization' => 'Bearer ' . $accessToken
+        );
+    }
+
+    /**
+     * GuzzleHTTP Client Implementation to use for HTTP requests
+     * @var Client
+     */
+    private $client;
 
     /**
      * ApiKey for the application
      * @var string
      */
-    protected $apiKey;
+    private $apiKey;
 
     /**
      * Constructor with the option to to supply an alternative rest client to be used
      * @param string $apiKey - Constant Contact API Key
-     * @param RestClientInterface $restClient - RestClientInterface implementation to be used in the service
      */
-    public function __construct($apiKey, $restClient = null)
+    public function __construct($apiKey)
     {
         $this->apiKey = $apiKey;
-
-        if (is_null($restClient)) {
-            $this->restClient = new RestClient();
-        } else {
-            $this->restClient = $restClient;
-        }
-    }
-
-    /**
-     * Build a URL from a base url and optional array of query parameters to append to the url. URL query parameters
-     * should not be URL encoded and this method will handle that.
-     * @param $url
-     * @param array $queryParams
-     * @return string
-     */
-    public function buildUrl($url, array $queryParams = null)
-    {
-        $keyArr = array('api_key' => $this->apiKey);
-        if ($queryParams) {
-            $params = array_merge($keyArr, $queryParams);
-        } else {
-            $params = $keyArr;
-        }
-
-        return $url . '?' . http_build_query($params, '', '&');
+        $this->client = new Client();
     }
 
     /**
      * Get the rest client being used by the service
-     * @return RestClientInterface - RestClientInterface implementation being used
+     * @return Client - GuzzleHTTP Client implementation being used
      */
-    public function getRestClient()
+    protected function getClient()
     {
-        return $this->restClient;
+        return $this->client;
     }
 
-    public function setRestClient(RestClientInterface $restClient)
-    {
-        $this->restClient = $restClient;
+    protected function createBaseRequest($accessToken, $method, $baseUrl) {
+        $request = $this->client->createRequest($method, $baseUrl);
+        $request->getQuery()->set("api_key", $this->apiKey);
+        $request->setHeaders($this->getHeaders($accessToken));
+        return $request;
     }
-
 
     /**
-     * Helper function to return required headers for making an http request with constant contact
-     * @param $accessToken - OAuth2 access token to be placed into the Authorization header
-     * @return array - authorization headers
+     * Turns a ClientException into a CtctException - like magic.
+     * @param ClientException $exception - Guzzle ClientException
+     * @return CtctException
      */
-    protected static function getHeaders($accessToken)
+    protected function convertException($exception)
     {
-        return array(
-            'Content-Type: application/json',
-            'Accept: application/json',
-            'Authorization: Bearer ' . $accessToken
-        );
+        $ctctException = new CtctException($exception->getResponse()->getReasonPhrase(), $exception->getCode());
+        $ctctException->setUrl($exception->getResponse()->getEffectiveUrl());
+        $ctctException->setErrors(json_decode($exception->getResponse()->getBody()->getContents()));
+        return $ctctException;
     }
 }
