@@ -1,9 +1,12 @@
 <?php
 namespace Ctct\Services;
 
+use Ctct\Exceptions\CtctException;
 use Ctct\Util\Config;
 use Ctct\Components\EmailMarketing\Schedule;
 use Ctct\Components\EmailMarketing\TestSend;
+use GuzzleHttp\Exception\ClientException;
+use GuzzleHttp\Stream\Stream;
 
 /**
  * Performs all actions pertaining to scheduling Constant Contact Campaigns
@@ -19,14 +22,23 @@ class CampaignScheduleService extends BaseService
      * @param int $campaignId - Campaign id to be scheduled
      * @param Schedule $schedule - Schedule to be created
      * @return Schedule
+     * @throws CtctException
      */
     public function addSchedule($accessToken, $campaignId, Schedule $schedule)
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_schedules'), $campaignId);
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->post($url, parent::getHeaders($accessToken), $schedule->toJson());
-        return Schedule::create(json_decode($response->body, true));
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_schedules'), $campaignId);
+
+        $request = parent::createBaseRequest($accessToken, 'POST', $baseUrl);
+        $stream = Stream::factory(json_encode($schedule));
+        $request->setBody($stream);
+
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        return Schedule::create($response->json());
     }
 
     /**
@@ -34,20 +46,24 @@ class CampaignScheduleService extends BaseService
      * @param string $accessToken - Constant Contact OAuth2 access token
      * @param int $campaignId - Campaign id to be scheduled
      * @return array
+     * @throws CtctException
      */
     public function getSchedules($accessToken, $campaignId)
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_schedules'), $campaignId);
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_schedules'), $campaignId);
 
-        $schedules = array();
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
 
-        foreach (json_decode($response->body, true) as $schedule) {
-            $schedules[] = Schedule::create($schedule);
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
         }
 
+        $schedules = array();
+        foreach ($response->json() as $schedule) {
+            $schedules[] = Schedule::create($schedule);
+        }
         return $schedules;
     }
 
@@ -57,14 +73,21 @@ class CampaignScheduleService extends BaseService
      * @param int $campaignId - Campaign id to be get a schedule for
      * @param int $scheduleId - Schedule id to retrieve
      * @return Schedule
+     * @throws CtctException
      */
     public function getSchedule($accessToken, $campaignId, $scheduleId)
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_schedule'), $campaignId, $scheduleId);
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
-        return Schedule::create(json_decode($response->body, true));
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_schedule'), $campaignId, $scheduleId);
+
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
+
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        return Schedule::create($response->json());
     }
 
     /**
@@ -73,45 +96,70 @@ class CampaignScheduleService extends BaseService
      * @param int $campaignId - Campaign id to be scheduled
      * @param Schedule $schedule - Schedule to retrieve
      * @return Schedule
+     * @throws CtctException
      */
     public function updateSchedule($accessToken, $campaignId, Schedule $schedule)
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_schedule'), $campaignId, $schedule->id);
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->put($url, parent::getHeaders($accessToken), $schedule->toJson());
-        return Schedule::create(json_decode($response->body, true));
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_schedule'), $campaignId, $schedule->id);
+
+        $request = parent::createBaseRequest($accessToken, 'PUT', $baseUrl);
+        $stream = Stream::factory(json_encode($schedule));
+        $request->setBody($stream);
+
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        return Schedule::create($response->json());
     }
 
     /**
-     * Get a specific schedule for a campaign
+     * Delete a specific schedule for a campaign
      * @param string $accessToken - Constant Contact OAuth2 access token
      * @param int $campaignId - Campaign id
      * @param int $scheduleId - Schedule id to delete
-     * @return Schedule
+     * @return True if successful
+     * @throws CtctException
      */
     public function deleteSchedule($accessToken, $campaignId, $scheduleId)
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_schedule'), $campaignId, $scheduleId);
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->delete($url, parent::getHeaders($accessToken));
-        return ($response->info['http_code'] == 204) ? true : false;
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_schedule'), $campaignId, $scheduleId);
+
+        $request = parent::createBaseRequest($accessToken, 'DELETE', $baseUrl);
+
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        return ($response->getStatusCode() == 204) ? true : false;
     }
 
     /**
      * Send a test send of a campaign
      * @param string $accessToken - Constant Contact OAuth2 access token
      * @param int $campaignId - Id of campaign to send test of
-     * @param TestSend $test_send - Test send details
+     * @param TestSend $testSend - Test send details
      * @return TestSend
+     * @throws CtctException
      */
-    public function sendTest($accessToken, $campaignId, TestSend $test_send)
+    public function sendTest($accessToken, $campaignId, TestSend $testSend)
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_test_sends'), $campaignId);
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->post($url, parent::getHeaders($accessToken), $test_send->toJson());
-        return TestSend::create(json_decode($response->body, true));
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_test_sends'), $campaignId);
+
+        $request = parent::createBaseRequest($accessToken, 'POST', $baseUrl);
+        $stream = Stream::factory(json_encode($testSend));
+        $request->setBody($stream);
+
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        return TestSend::create($response->json());
     }
 }
