@@ -1,6 +1,7 @@
 <?php
 namespace Ctct\Services;
 
+use Ctct\Exceptions\CtctException;
 use Ctct\Util\Config;
 use Ctct\Components\Tracking\BounceActivity;
 use Ctct\Components\Tracking\TrackingActivity;
@@ -11,6 +12,7 @@ use Ctct\Components\Tracking\UnsubscribeActivity;
 use Ctct\Components\Tracking\SendActivity;
 use Ctct\Components\Tracking\TrackingSummary;
 use Ctct\Components\ResultSet;
+use GuzzleHttp\Exception\ClientException;
 
 /**
  * Performs all actions pertaining to Constant Contact Campaign Tracking
@@ -24,19 +26,34 @@ class CampaignTrackingService extends BaseService
     /**
      * Get a result set of bounces for a given campaign
      * @param string $accessToken - Constant Contact OAuth2 access token
-     * @param string $campaign_id - Campaign id
-     * @param array $params - query parameters to be appended to the request
+     * @param string $campaignId - Campaign id
+     * @param array $params - associative array of query parameters and values to append to the request.
+     *      Allowed parameters include:
+     *      limit - Specifies the number of results displayed per page of output, from 1 - 500, default = 50.
+     *      created_since - Used to retrieve a list of events since the date and time specified (in ISO-8601 format).
+     *      next - the next link returned from a previous paginated call. May only be used by itself.
      * @return ResultSet - Containing a results array of {@link BounceActivity}
+     * @throws CtctException
      */
-    public function getBounces($accessToken, $campaign_id, array $params = array())
+    public function getBounces($accessToken, $campaignId, Array $params = array())
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_tracking_bounces'), $campaign_id);
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_tracking_bounces'), $campaignId);
 
-        $url = $this->buildUrl($baseUrl, $params);
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
+        if ($params) {
+            $query = $request->getQuery();
+            foreach ($params as $name => $value) {
+                $query->add($name, $value);
+            }
+        }
 
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
-        $body = json_decode($response->body, true);
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        $body = $response->json();
         $bounces = array();
         foreach ($body['results'] as $bounceActivity) {
             $bounces[] = BounceActivity::create($bounceActivity);
@@ -48,19 +65,33 @@ class CampaignTrackingService extends BaseService
      * Get clicks for a given campaign
      * @param string $accessToken - Constant Contact OAuth2 access token
      * @param string $campaignId - Campaign id
-     * @param array $params - query params to be appended to request
+     * @param array $params - associative array of query parameters and values to append to the request.
+     *      Allowed parameters include:
+     *      limit - Specifies the number of results displayed per page of output, from 1 - 500, default = 50.
+     *      created_since - Used to retrieve a list of events since the date and time specified (in ISO-8601 format).
+     *      next - the next link returned from a previous paginated call. May only be used by itself.
      * @return ResultSet - Containing a results array of {@link ClickActivity}
+     * @throws CtctException
      */
-    public function getClicks($accessToken, $campaignId, array $params = array())
+    public function getClicks($accessToken, $campaignId, Array $params = array())
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_tracking_clicks'), $campaignId);
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_tracking_clicks'), $campaignId);
 
-        $url = $this->buildUrl($baseUrl, $params);
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
+        if ($params) {
+            $query = $request->getQuery();
+            foreach ($params as $name => $value) {
+                $query->add($name, $value);
+            }
+        }
 
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
-        $body = json_decode($response->body, true);
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
 
+        $body = $response->json();
         $clicks = array();
         foreach ($body['results'] as $click_activity) {
             $clicks[] = ClickActivity::create($click_activity);
@@ -73,18 +104,33 @@ class CampaignTrackingService extends BaseService
      * Get forwards for a given campaign
      * @param string $accessToken - Constant Contact OAuth2 access token
      * @param string $campaignId - Campaign id
-     * @param array $params - query param to be appended to request
+     * @param array $params - associative array of query parameters and values to append to the request.
+     *      Allowed parameters include:
+     *      limit - Specifies the number of results displayed per page of output, from 1 - 500, default = 50.
+     *      created_since - Used to retrieve a list of events since the date and time specified (in ISO-8601 format).
+     *      next - the next link returned from a previous paginated call. May only be used by itself.
      * @return ResultSet - Containing a results array of {@link ForwardActivity}
+     * @throws CtctException
      */
-    public function getForwards($accessToken, $campaignId, array $params = array())
+    public function getForwards($accessToken, $campaignId, Array $params = array())
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_tracking_forwards'), $campaignId);
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_tracking_forwards'), $campaignId);
 
-        $url = $this->buildUrl($baseUrl, $params);
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
+        if ($params) {
+            $query = $request->getQuery();
+            foreach ($params as $name => $value) {
+                $query->add($name, $value);
+            }
+        }
 
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
-        $body = json_decode($response->body, true);
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        $body = $response->json();
         $forwards = array();
         foreach ($body['results'] as $forward_activity) {
             $forwards[] = ForwardActivity::create($forward_activity);
@@ -96,19 +142,34 @@ class CampaignTrackingService extends BaseService
     /**
      * Get opens for a given campaign
      * @param string $accessToken - Constant Contact OAuth2 access token
-     * @param string $campaign_id - Campaign id
-     * @param array $params - query params to be appended to request
+     * @param string $campaignId - Campaign id
+     * @param array $params - associative array of query parameters and values to append to the request.
+     *      Allowed parameters include:
+     *      limit - Specifies the number of results displayed per page of output, from 1 - 500, default = 50.
+     *      created_since - Used to retrieve a list of events since the date and time specified (in ISO-8601 format).
+     *      next - the next link returned from a previous paginated call. May only be used by itself.
      * @return ResultSet - Containing a results array of {@link OpenActivity}
+     * @throws CtctException
      */
-    public function getOpens($accessToken, $campaign_id, array $params = array())
+    public function getOpens($accessToken, $campaignId, Array $params = array())
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_tracking_opens'), $campaign_id);
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_tracking_opens'), $campaignId);
 
-        $url = $this->buildUrl($baseUrl, $params);
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
+        if ($params) {
+            $query = $request->getQuery();
+            foreach ($params as $name => $value) {
+                $query->add($name, $value);
+            }
+        }
 
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
-        $body = json_decode($response->body, true);
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        $body = $response->json();
         $opens = array();
         foreach ($body['results'] as $open_activity) {
             $opens[] = OpenActivity::create($open_activity);
@@ -121,18 +182,33 @@ class CampaignTrackingService extends BaseService
      * Get sends for a given campaign
      * @param string $accessToken - Constant Contact OAuth2 access token
      * @param string $campaignId - Campaign id
-     * @param Array $params - query params to be appended to request
+     * @param Array $params = array() - associative array of query parameters and values to append to the request.
+     *      Allowed parameters include:
+     *      limit - Specifies the number of results displayed per page of output, from 1 - 500, default = 50.
+     *      created_since - Used to retrieve a list of events since the date and time specified (in ISO-8601 format).
+     *      next - the next link returned from a previous paginated call. May only be used by itself.
      * @return TrackingActivity - Containing a results array of {@link SendActivity}
+     * @throws CtctException
      */
-    public function getSends($accessToken, $campaignId, array $params = array())
+    public function getSends($accessToken, $campaignId, Array $params = array())
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_tracking_sends'), $campaignId);
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_tracking_sends'), $campaignId);
 
-        $url = $this->buildUrl($baseUrl, $params);
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
+        if ($params) {
+            $query = $request->getQuery();
+            foreach ($params as $name => $value) {
+                $query->add($name, $value);
+            }
+        }
 
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
-        $body = json_decode($response->body, true);
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        $body = $response->json();
         $sends = array();
         foreach ($body['results'] as $send_activity) {
             $sends[] = SendActivity::create($send_activity);
@@ -145,18 +221,33 @@ class CampaignTrackingService extends BaseService
      * Get unsubscribes for a given campaign
      * @param string $accessToken - Constant Contact OAuth2 access token
      * @param string $campaignId - Campaign id
-     * @param array $params - query params to be appended to request
+     * @param array $params - associative array of query parameters and values to append to the request.
+     *      Allowed parameters include:
+     *      limit - Specifies the number of results displayed per page of output, from 1 - 500, default = 50.
+     *      created_since - Used to retrieve a list of events since the date and time specified (in ISO-8601 format).
+     *      next - the next link returned from a previous paginated call. May only be used by itself.
      * @return ResultSet - Containing a results array of {@link UnsubscribeActivity}
+     * @throws CtctException
      */
-    public function getUnsubscribes($accessToken, $campaignId, array $params = array())
+    public function getUnsubscribes($accessToken, $campaignId, Array $params = array())
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_tracking_unsubscribes'), $campaignId);
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_tracking_unsubscribes'), $campaignId);
 
-        $url = $this->buildUrl($baseUrl, $params);
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
+        if ($params) {
+            $query = $request->getQuery();
+            foreach ($params as $name => $value) {
+                $query->add($name, $value);
+            }
+        }
 
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
-        $body = json_decode($response->body, true);
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        $body = $response->json();
         $optOuts = array();
         foreach ($body['results'] as $opt_out_activity) {
             $optOuts[] = UnsubscribeActivity::create($opt_out_activity);
@@ -168,15 +259,22 @@ class CampaignTrackingService extends BaseService
     /**
      * Get a summary of reporting data for a given campaign
      * @param string $accessToken - Constant Contact OAuth2 access token
-     * @param int $campaign_id - Campaign id
+     * @param int $campaignId - Campaign id
      * @return TrackingSummary
+     * @throws CtctException
      */
-    public function getSummary($accessToken, $campaign_id)
+    public function getSummary($accessToken, $campaignId)
     {
-        $baseUrl = Config::get('endpoints.base_url') .
-            sprintf(Config::get('endpoints.campaign_tracking_summary'), $campaign_id);
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
-        return TrackingSummary::create(json_decode($response->body, true));
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.campaign_tracking_summary'), $campaignId);
+
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
+
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        return TrackingSummary::create($response->json());
     }
 }
