@@ -1,10 +1,11 @@
 <?php
 namespace Ctct\Services;
 
+use Ctct\Exceptions\CtctException;
 use Ctct\Util\Config;
 use Ctct\Components\Contacts\ContactList;
-use Ctct\Components\Contacts\Contact;
-use Ctct\Components\ResultSet;
+use GuzzleHttp\Exception\ClientException;
+use GuzzleHttp\Stream\Stream;
 
 /**
  * Performs all actions pertaining to Constant Contact Lists
@@ -17,19 +18,35 @@ class ListService extends BaseService
     /**
      * Get lists within an account
      * @param $accessToken - Constant Contact OAuth2 access token
-     * @param array $params - array of query parameters to be appended to the request
+     * @param array $params - associative array of query parameters and values to append to the request.
+     *      Allowed parameters include:
+     *      modified_since - ISO-8601 formatted timestamp.
      * @return Array - ContactLists
+     * @throws CtctException
      */
-    public function getLists($accessToken, array $params = array())
+    public function getLists($accessToken, Array $params = array())
     {
         $baseUrl = Config::get('endpoints.base_url') . Config::get('endpoints.lists');
-        $url = $this->buildUrl($baseUrl, $params);
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
+
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
+        if ($params) {
+            $query = $request->getQuery();
+            foreach ($params as $name => $value) {
+                $query->add($name, $value);
+            }
+        }
+
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
 
         $lists = array();
-        foreach (json_decode($response->body, true) as $contact) {
+        foreach ($response->json() as $contact) {
             $lists[] = ContactList::create($contact);
         }
+
         return $lists;
     }
 
@@ -38,13 +55,23 @@ class ListService extends BaseService
      * @param string $accessToken - Constant Contact OAuth2 access token
      * @param ContactList $list
      * @return ContactList
+     * @throws CtctException
      */
     public function addList($accessToken, ContactList $list)
     {
         $baseUrl = Config::get('endpoints.base_url') . Config::get('endpoints.lists');
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->post($url, parent::getHeaders($accessToken), $list->toJson());
-        return ContactList::create(json_decode($response->body, true));
+
+        $request = parent::createBaseRequest($accessToken, 'POST', $baseUrl);
+        $stream = Stream::factory(json_encode($list));
+        $request->setBody($stream);
+
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        return ContactList::create($response->json());
     }
 
     /**
@@ -52,61 +79,66 @@ class ListService extends BaseService
      * @param string $accessToken - Constant Contact OAuth2 access token
      * @param ContactList $list - ContactList to be updated
      * @return ContactList
+     * @throws CtctException
      */
     public function updateList($accessToken, ContactList $list)
     {
         $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.list'), $list->id);
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->put($url, parent::getHeaders($accessToken), $list->toJson());
-        return ContactList::create(json_decode($response->body, true));
+
+        $request = parent::createBaseRequest($accessToken, 'PUT', $baseUrl);
+        $stream = Stream::factory(json_encode($list));
+        $request->setBody($stream);
+
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        return ContactList::create($response->json());
     }
 
     /**
      * Delete a Contact List
      * @param string $accessToken - Constant Contact OAuth2 access token
-     * @param $list_id - list id
+     * @param $listId - list id
      * @return ContactList
+     * @throws CtctException
      */
-    public function deleteList($accessToken, $list_id)
+    public function deleteList($accessToken, $listId)
     {
-        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.list'), $list_id);
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->delete($url, parent::getHeaders($accessToken));
-        return ($response->info['http_code'] == 204) ? true : false;
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.list'), $listId);
+
+        $request = parent::createBaseRequest($accessToken, 'DELETE', $baseUrl);
+
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
+        }
+
+        return ($response->getStatusCode() == 204) ? true : false;
     }
 
     /**
      * Get an individual contact list
      * @param $accessToken - Constant Contact OAuth2 access token
-     * @param $list_id - list id
+     * @param $listId - list id
      * @return ContactList
+     * @throws CtctException
      */
-    public function getList($accessToken, $list_id)
+    public function getList($accessToken, $listId)
     {
-        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.list'), $list_id);
-        $url = $this->buildUrl($baseUrl);
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
-        return ContactList::create(json_decode($response->body, true));
-    }
+        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.list'), $listId);
 
-    /**
-     * Get all contacts from an individual list
-     * @param string $accessToken - Constant Contact OAuth2 access token
-     * @param string $list_id - list id to retrieve contacts for
-     * @param array $params - query params to attach to request
-     * @return ResultSet
-     */
-    public function getContactsFromList($accessToken, $list_id, $params = null)
-    {
-        $baseUrl = Config::get('endpoints.base_url') . sprintf(Config::get('endpoints.list_contacts'), $list_id);
-        $url = $this->buildUrl($baseUrl, $params);
+        $request = parent::createBaseRequest($accessToken, 'GET', $baseUrl);
 
-        $response = parent::getRestClient()->get($url, parent::getHeaders($accessToken));
-        $body = json_decode($response->body, true);
-        $contacts = array();
-        foreach ($body['results'] as $contact) {
-            $contacts[] = Contact::create($contact);
+        try {
+            $response = parent::getClient()->send($request);
+        } catch (ClientException $e) {
+            throw parent::convertException($e);
         }
-        return new ResultSet($contacts, $body['meta']);
+
+        return ContactList::create($response->json());
     }
 }
