@@ -1,57 +1,62 @@
 <?php
 
-use Ctct\Services\AccountService;
-use Ctct\Util\CurlResponse;
+use Ctct\Components\Account\AccountInfo;
+use Ctct\Components\Account\VerifiedEmailAddress;
+use GuzzleHttp\Client;
+use GuzzleHttp\Subscriber\Mock;
+use GuzzleHttp\Stream\Stream;
+use GuzzleHttp\Message\Response;
 
 class AccountServiceUnitTest extends PHPUnit_Framework_TestCase
 {
+    /**
+     * @var Client
+     */
+    private static $client;
 
-    private $restClient;
-    private $accountService;
-
-    public function setUp()
+    public static function setUpBeforeClass()
     {
-        $this->restClient = $this->getMock('Ctct\Util\RestClientInterface');
-        $this->accountService = new AccountService("apikey", $this->restClient);
+        self::$client = new Client();
+        $verifiedAddressStream = Stream::factory(JsonLoader::getVerifiedAddressesJson());
+        $accountInfoStream = Stream::factory(JsonLoader::getAccountInfoJson());
+        $mock = new Mock([
+            new Response(200, array(), $verifiedAddressStream),
+            new Response(200, array(), $accountInfoStream)
+        ]);
+        self::$client->getEmitter()->attach($mock);
     }
 
-    public function testGetVerifiedAddresses()
-    {
-        $curlResponse = CurlResponse::create(JsonLoader::getVerifiedAddressesJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+    public function testGetVerifiedAddresses() {
+        $response = self::$client->get('/');
+        $verifiedAddresses = array();
+        foreach ($response->json() as $verifiedAddress) {
+            $verifiedAddresses[] = VerifiedEmailAddress::create($verifiedAddress);
+        }
 
-        $response = $this->accountService->getVerifiedEmailAddresses("accessToken", array());
-
-        $this->assertInstanceOf('Ctct\Components\Account\VerifiedEmailAddress', $response[0]);
-        $this->assertEquals("test123@roving.com", $response[0]->email_address);
-        $this->assertEquals("CONFIRMED", $response[0]->status);
+        foreach ($verifiedAddresses as $verifiedAddress) {
+            $this->assertInstanceOf('Ctct\Components\Account\VerifiedEmailAddress', $verifiedAddress);
+            $this->assertEquals("test123@roving.com", $verifiedAddress->email_address);
+            $this->assertEquals("CONFIRMED", $verifiedAddress->status);
+        }
     }
 
     public function testGetAccountInfo()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getAccountInfoJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
+        $result = AccountInfo::create($response->json());
 
-        $response = $this->accountService->getAccountInfo("accessToken");
-
-        $this->assertInstanceOf('Ctct\Components\Account\AccountInfo', $response);
-        $this->assertEquals("http://www.example.com", $response->website);
-        $this->assertEquals("My Company", $response->organization_name);
-        $this->assertEquals("http://www.example.com", $response->website);
-        $this->assertEquals("My Company", $response->organization_name);
-        $this->assertEquals("US/Eastern", $response->time_zone);
-        $this->assertEquals("Mary Jane", $response->first_name);
-        $this->assertEquals("Doe", $response->last_name);
-        $this->assertEquals("mjdoe@example.com", $response->email);
-        $this->assertEquals("5555555555", $response->phone);
-        $this->assertEquals("https://ih.constantcontact.com/fs137/1100371573368/img/90.jpg", $response->company_logo);
-        $this->assertEquals("US", $response->country_code);
-        $this->assertEquals("MA", $response->state_code);
+        $this->assertInstanceOf('Ctct\Components\Account\AccountInfo', $result);
+        $this->assertEquals("http://www.example.com", $result->website);
+        $this->assertEquals("My Company", $result->organization_name);
+        $this->assertEquals("http://www.example.com", $result->website);
+        $this->assertEquals("My Company", $result->organization_name);
+        $this->assertEquals("US/Eastern", $result->time_zone);
+        $this->assertEquals("Mary Jane", $result->first_name);
+        $this->assertEquals("Doe", $result->last_name);
+        $this->assertEquals("mjdoe@example.com", $result->email);
+        $this->assertEquals("5555555555", $result->phone);
+        $this->assertEquals("https://ih.constantcontact.com/fs137/1100371573368/img/90.jpg", $result->company_logo);
+        $this->assertEquals("US", $result->country_code);
+        $this->assertEquals("MA", $result->state_code);
     }
 }
