@@ -1,49 +1,63 @@
 <?php
 
-use Ctct\Services\CampaignScheduleService;
-use Ctct\Util\RestClient;
-use Ctct\Util\CurlResponse;
 use Ctct\Components\EmailMarketing\Schedule;
 use Ctct\Components\EmailMarketing\TestSend;
 
+use GuzzleHttp\Client;
+use GuzzleHttp\Exception\ClientException;
+use GuzzleHttp\Subscriber\Mock;
+use GuzzleHttp\Stream\Stream;
+use GuzzleHttp\Message\Response;
+
 class CampaignScheduleServiceUnitTest extends PHPUnit_Framework_TestCase
 {
+    /**
+     * @var Client
+     */
+    private static $client;
 
-    private $restClient;
-    private $campaignScheduleService;
-
-    public function setUp()
+    public static function setUpBeforeClass()
     {
-        $this->restClient = $this->getMock('Ctct\Util\RestClientInterface');
-        $this->campaignScheduleService = new CampaignScheduleService("apikey", $this->restClient);
+        self::$client = new Client();
+        $schedulesStream = Stream::factory(JsonLoader::getCampaignSchedulesJson());
+        $scheduleStream = Stream::factory(JsonLoader::getCampaignScheduleJson());
+        $testSendStream = Stream::factory(JsonLoader::getTestSendJson());
+        $mock = new Mock([
+            new Response(200, array(), $schedulesStream),
+            new Response(200, array(), $scheduleStream),
+            new Response(201, array(), $scheduleStream),
+            new Response(200, array(), $scheduleStream),
+            new Response(204, array()),
+            new Response(400, array()),
+            new Response(200, array(), $testSendStream)
+        ]);
+        self::$client->getEmitter()->attach($mock);
     }
 
     public function testGetSchedules()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getCampaignSchedulesJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $schedules = $this->campaignScheduleService->getSchedules('access_token', 9100367935463);
-        $this->assertInstanceOf('Ctct\Components\EmailMarketing\Schedule', $schedules[0]);
-        $this->assertEquals(1, $schedules[0]->id);
-        $this->assertEquals("2012-12-16T11:07:43.626Z", $schedules[0]->scheduled_date);
+        $schedules = array();
+        foreach ($response->json() as $responseSchedule) {
+            $schedules[] = $responseSchedule;
+        }
 
-        $this->assertEquals(2, $schedules[1]->id);
-        $this->assertEquals("2012-12-17T11:08:00.000Z", $schedules[1]->scheduled_date);
+        $schedule1 = Schedule::create($schedules[0]);
+        $this->assertInstanceOf('Ctct\Components\EmailMarketing\Schedule', $schedule1);
+        $this->assertEquals(1, $schedule1->id);
+        $this->assertEquals("2012-12-16T11:07:43.626Z", $schedule1->scheduled_date);
+
+        $schedule2 = Schedule::create($schedules[1]);
+        $this->assertEquals(2, $schedule2->id);
+        $this->assertEquals("2012-12-17T11:08:00.000Z", $schedule2->scheduled_date);
     }
 
     public function testGetSchedule()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getCampaignScheduleJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $schedule = $this->campaignScheduleService->getSchedule('access_token', 9100367935463, 1);
+        $schedule = Schedule::create($response->json());
         $this->assertInstanceOf('Ctct\Components\EmailMarketing\Schedule', $schedule);
         $this->assertEquals(1, $schedule->id);
         $this->assertEquals("2012-12-16T11:07:43.626Z", $schedule->scheduled_date);
@@ -51,13 +65,9 @@ class CampaignScheduleServiceUnitTest extends PHPUnit_Framework_TestCase
 
     public function testAddSchedule()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getCampaignScheduleJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('post')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->post('/');
 
-        $createdSchedule = $this->campaignScheduleService->addSchedule('access_token', 9100367935463, new Schedule());
+        $createdSchedule = Schedule::create($response->json());
         $this->assertInstanceOf('Ctct\Components\EmailMarketing\Schedule', $createdSchedule);
         $this->assertEquals(1, $createdSchedule->id);
         $this->assertEquals("2012-12-16T11:07:43.626Z", $createdSchedule->scheduled_date);
@@ -65,18 +75,9 @@ class CampaignScheduleServiceUnitTest extends PHPUnit_Framework_TestCase
 
     public function testUpdateSchedule()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getCampaignScheduleJson(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('put')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->put('/');
 
-        $updatedSchedule = $this->campaignScheduleService->updateSchedule(
-            'access_token',
-            "9100367935463",
-            new Schedule()
-        );
-
+        $updatedSchedule = Schedule::create($response->json());
         $this->assertInstanceOf('Ctct\Components\EmailMarketing\Schedule', $updatedSchedule);
         $this->assertEquals(1, $updatedSchedule->id);
         $this->assertEquals("2012-12-16T11:07:43.626Z", $updatedSchedule->scheduled_date);
@@ -84,35 +85,26 @@ class CampaignScheduleServiceUnitTest extends PHPUnit_Framework_TestCase
 
     public function testDeleteSchedule()
     {
-        $curlResponse = CurlResponse::create("", array('http_code' => 204), null);
-        $this->restClient->expects($this->once())
-            ->method('delete')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->delete('/');
 
-        $this->assertTrue($this->campaignScheduleService->deleteSchedule('access_token', "9100367935463", 1));
+        $this->assertEquals(204, $response->getStatusCode());
     }
 
     public function testDeleteScheduleFailed()
     {
-        $curlResponse = CurlResponse::create("", array('http_code' => 400));
-        $this->restClient->expects($this->once())
-            ->method('delete')
-            ->with()
-            ->will($this->returnValue($curlResponse));
-
-        $this->assertFalse($this->campaignScheduleService->deleteSchedule('access_token', "9100367935463", 1));
+        try {
+            self::$client->delete('/');
+            $this->fail("Call did not fail");
+        } catch (ClientException $e) {
+            $this->assertEquals(400, $e->getCode());
+        }
     }
 
     public function testSendTest()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getTestSendJson(), array('http_code' => 201));
-        $this->restClient->expects($this->once())
-            ->method('post')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->post('/');
 
-        $testSend = $this->campaignScheduleService->sendTest('access_token', "9100367935463", new TestSend());
+        $testSend = TestSend::create($response->json());
         $this->assertInstanceOf('Ctct\Components\EmailMarketing\TestSend', $testSend);
         $this->assertEquals("HTML", $testSend->format);
         $this->assertEquals("oh hai there", $testSend->personal_message);
