@@ -1,185 +1,197 @@
 <?php
 
-use Ctct\Services\CampaignTrackingService;
-use Ctct\Util\RestClient;
-use Ctct\Util\CurlResponse;
+use Ctct\Components\ResultSet;
+use Ctct\Components\Tracking\BounceActivity;
+use Ctct\Components\Tracking\ClickActivity;
+use Ctct\Components\Tracking\ForwardActivity;
+use Ctct\Components\Tracking\OpenActivity;
+use Ctct\Components\Tracking\SendActivity;
+use Ctct\Components\Tracking\TrackingSummary;
+use Ctct\Components\Tracking\UnsubscribeActivity;
+
+use GuzzleHttp\Client;
+use GuzzleHttp\Subscriber\Mock;
+use GuzzleHttp\Stream\Stream;
+use GuzzleHttp\Message\Response;
 
 class CampaignTrackingServiceUnitTest extends PHPUnit_Framework_TestCase
 {
+    /**
+     * @var Client
+     */
+    private static $client;
 
-    private $restClient;
-    private $campaignTrackingService;
-
-    public function setUp()
+    public static function setUpBeforeClass()
     {
-        $this->restClient = $this->getMock('Ctct\Util\RestClientInterface');
-        $this->campaignTrackingService = new CampaignTrackingService("apikey", $this->restClient);
+        self::$client = new Client();
+        $bouncesStream = Stream::factory(JsonLoader::getBounces());
+        $clicksStream = Stream::factory(JsonLoader::getClicks());
+        $forwardsStream = Stream::factory(JsonLoader::getForwards());
+        $unsubscribesStream = Stream::factory(JsonLoader::getOptOuts());
+        $sendsStream = Stream::factory(JsonLoader::getSends());
+        $opensStream = Stream::factory(JsonLoader::getOpens());
+        $summaryStream = Stream::factory(JsonLoader::getSummary());
+        $mock = new Mock([
+            new Response(200, array(), $bouncesStream),
+            new Response(200, array(), $clicksStream),
+            new Response(200, array(), $forwardsStream),
+            new Response(200, array(), $unsubscribesStream),
+            new Response(200, array(), $sendsStream),
+            new Response(200, array(), $opensStream),
+            new Response(200, array(), $summaryStream)
+        ]);
+        self::$client->getEmitter()->attach($mock);
     }
 
     public function testGetBounces()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getBounces(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $resultSet = $this->campaignTrackingService->getBounces('access_token', "1100394165290", array('limit' => 2));
+        $responseJson = $response->json();
+        $resultSet = new ResultSet($responseJson['results'], $responseJson['meta']);
+
+        $bounceActivity = BounceActivity::create($resultSet->results[0]);
 
         $this->assertInstanceOf('Ctct\Components\ResultSet', $resultSet);
-        $this->assertInstanceOf('Ctct\Components\Tracking\BounceActivity', $resultSet->results[0]);
+        $this->assertInstanceOf('Ctct\Components\Tracking\BounceActivity', $bounceActivity);
         $this->assertEquals("bGltaXQ9MyZuZXh0PTEzNTQ4MTcyMTA0MzA", $resultSet->next);
-        $this->assertEquals("EMAIL_BOUNCE", $resultSet->results[0]->activity_type);
-        $this->assertEquals(1100394165290, $resultSet->results[0]->campaign_id);
-        $this->assertEquals("2", $resultSet->results[0]->contact_id);
+        $this->assertEquals("EMAIL_BOUNCE", $bounceActivity->activity_type);
+        $this->assertEquals(1100394165290, $bounceActivity->campaign_id);
+        $this->assertEquals("2", $bounceActivity->contact_id);
         $this->assertEquals(
             "wizpie9dedde9dd27644bdb3d9be134b7294f71354817123188401000@snoopy.roving.com",
-            $resultSet->results[0]->email_address
+            $bounceActivity->email_address
         );
 
-        $this->assertEquals("B", $resultSet->results[0]->bounce_code);
-        $this->assertEquals("Non-existent address", $resultSet->results[0]->bounce_description);
-        $this->assertEquals("", $resultSet->results[0]->bounce_message);
-        $this->assertEquals("2012-12-06T13:05:24.844Z", $resultSet->results[0]->bounce_date);
+        $this->assertEquals("B", $bounceActivity->bounce_code);
+        $this->assertEquals("Non-existent address", $bounceActivity->bounce_description);
+        $this->assertEquals("", $bounceActivity->bounce_message);
+        $this->assertEquals("2012-12-06T13:05:24.844Z", $bounceActivity->bounce_date);
     }
 
     public function testGetClicks()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getClicks(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $resultSet = $this->campaignTrackingService->getClicks('access_token', "1100394165290", array('limit' => 2));
+        $responseJson = $response->json();
+        $resultSet = new ResultSet($responseJson['results'], $responseJson['meta']);
+
+        $clickActivity = ClickActivity::create($resultSet->results[0]);
 
         $this->assertInstanceOf('Ctct\Components\ResultSet', $resultSet);
-        $this->assertInstanceOf('Ctct\Components\Tracking\ClickActivity', $resultSet->results[0]);
+        $this->assertInstanceOf('Ctct\Components\Tracking\ClickActivity', $clickActivity);
         $this->assertEquals("bGltaXQ9MyZuZXh0PTEzNTQ4MTcyMTA0MzA", $resultSet->next);
-        $this->assertEquals("EMAIL_CLICK", $resultSet->results[0]->activity_type);
-        $this->assertEquals(1100394165290, $resultSet->results[0]->campaign_id);
-        $this->assertEquals("69", $resultSet->results[0]->contact_id);
+        $this->assertEquals("EMAIL_CLICK", $clickActivity->activity_type);
+        $this->assertEquals(1100394165290, $clickActivity->campaign_id);
+        $this->assertEquals("69", $clickActivity->contact_id);
         $this->assertEquals(
             "wizpie00375ca0a11346a89aea4b8f5991d0d91354817217769892000@snoopy.roving.com",
-            $resultSet->results[0]->email_address
+            $clickActivity->email_address
         );
 
-        $this->assertEquals(0, $resultSet->results[0]->link_id);
-        $this->assertEquals("2012-12-06T13:07:01.701Z", $resultSet->results[0]->click_date);
+        $this->assertEquals(0, $clickActivity->link_id);
+        $this->assertEquals("2012-12-06T13:07:01.701Z", $clickActivity->click_date);
     }
 
     public function testGetForwards()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getForwards(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $resultSet = $this->campaignTrackingService->getForwards('access_token', "1100394165290", array('limit' => 2));
+        $responseJson = $response->json();
+        $resultSet = new ResultSet($responseJson['results'], $responseJson['meta']);
+
+        $forwardActivity = ForwardActivity::create($resultSet->results[0]);
 
         $this->assertInstanceOf('Ctct\Components\ResultSet', $resultSet);
-        $this->assertInstanceOf('Ctct\Components\Tracking\ForwardActivity', $resultSet->results[0]);
+        $this->assertInstanceOf('Ctct\Components\Tracking\ForwardActivity', $forwardActivity);
         $this->assertEquals("bGltaXQ9MyZuZXh0PTEzNTQ4MTcyMTA0MzA", $resultSet->next);
-        $this->assertEquals("EMAIL_FORWARD", $resultSet->results[0]->activity_type);
-        $this->assertEquals(1100394165290, $resultSet->results[0]->campaign_id);
-        $this->assertEquals("74", $resultSet->results[0]->contact_id);
+        $this->assertEquals("EMAIL_FORWARD", $forwardActivity->activity_type);
+        $this->assertEquals(1100394165290, $forwardActivity->campaign_id);
+        $this->assertEquals("74", $forwardActivity->contact_id);
         $this->assertEquals(
             "wizpie2ca3455df5c34a26806f519f01f8a22e1354817223114268000@snoopy.roving.com",
-            $resultSet->results[0]->email_address
+            $forwardActivity->email_address
         );
 
-        $this->assertEquals("2012-12-06T13:07:06.810Z", $resultSet->results[0]->forward_date);
+        $this->assertEquals("2012-12-06T13:07:06.810Z", $forwardActivity->forward_date);
     }
 
     public function testGetUnsubscribes()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getOptOuts(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $resultSet = $this->campaignTrackingService->getUnsubscribes(
-            'access_token',
-            "1100394165290",
-            array('limit' => 2)
-        );
+        $responseJson = $response->json();
+        $resultSet = new ResultSet($responseJson['results'], $responseJson['meta']);
+
+        $unsubscribeActivity = UnsubscribeActivity::create($resultSet->results[0]);
 
         $this->assertInstanceOf('Ctct\Components\ResultSet', $resultSet);
-        $this->assertInstanceOf('Ctct\Components\Tracking\UnsubscribeActivity', $resultSet->results[0]);
+        $this->assertInstanceOf('Ctct\Components\Tracking\UnsubscribeActivity', $unsubscribeActivity);
         $this->assertEquals("bGltaXQ9MyZuZXh0PTEzNTQ4MTcyMTA0MzA", $resultSet->next);
-        $this->assertEquals("EMAIL_UNSUBSCRIBE", $resultSet->results[0]->activity_type);
-        $this->assertEquals(1100394165290, $resultSet->results[0]->campaign_id);
-        $this->assertEquals("58", $resultSet->results[0]->contact_id);
+        $this->assertEquals("EMAIL_UNSUBSCRIBE", $unsubscribeActivity->activity_type);
+        $this->assertEquals(1100394165290, $unsubscribeActivity->campaign_id);
+        $this->assertEquals("58", $unsubscribeActivity->contact_id);
         $this->assertEquals(
             "wizpieabd7817c1d0d4f08bb05f16f6681221c1354817211855027000@snoopy.roving.com",
-            $resultSet->results[0]->email_address
+            $unsubscribeActivity->email_address
         );
 
-        $this->assertEquals("2012-12-06T13:06:53.440Z", $resultSet->results[0]->unsubscribe_date);
-        $this->assertEquals("ACTION_BY_CUSTOMER", $resultSet->results[0]->unsubscribe_source);
-        $this->assertEquals("", $resultSet->results[0]->unsubscribe_reason);
+        $this->assertEquals("2012-12-06T13:06:53.440Z", $unsubscribeActivity->unsubscribe_date);
+        $this->assertEquals("ACTION_BY_CUSTOMER", $unsubscribeActivity->unsubscribe_source);
+        $this->assertEquals("", $unsubscribeActivity->unsubscribe_reason);
     }
 
     public function testGetSends()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getSends(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $resultSet = $this->campaignTrackingService->getSends('access_token', "1100394165290", array('limit' => 2));
+        $responseJson = $response->json();
+        $resultSet = new ResultSet($responseJson['results'], $responseJson['meta']);
+
+        $sendActivity = SendActivity::create($resultSet->results[0]);
 
         $this->assertInstanceOf('Ctct\Components\ResultSet', $resultSet);
-        $this->assertInstanceOf('Ctct\Components\Tracking\SendActivity', $resultSet->results[0]);
+        $this->assertInstanceOf('Ctct\Components\Tracking\SendActivity', $sendActivity);
         $this->assertEquals("bGltaXQ9MyZuZXh0PTEzNTQ4MTcyMTA0MzA", $resultSet->next);
-        $this->assertEquals("EMAIL_SEND", $resultSet->results[0]->activity_type);
-        $this->assertEquals(1100394165290, $resultSet->results[0]->campaign_id);
-        $this->assertEquals("55", $resultSet->results[0]->contact_id);
+        $this->assertEquals("EMAIL_SEND", $sendActivity->activity_type);
+        $this->assertEquals(1100394165290, $sendActivity->campaign_id);
+        $this->assertEquals("55", $sendActivity->contact_id);
         $this->assertEquals(
             "wizpiea298d1c2500b4f2d8294300de4b29fe31354817207606824000@snoopy.roving.com",
-            $resultSet->results[0]->email_address
+            $sendActivity->email_address
         );
 
-        $this->assertEquals("2012-12-06T18:06:50.650Z", $resultSet->results[0]->send_date);
+        $this->assertEquals("2012-12-06T18:06:50.650Z", $sendActivity->send_date);
     }
 
     public function testGetOpens()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getOpens(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $resultSet = $this->campaignTrackingService->getOpens('access_token', "1100394165290", array('limit' => 2));
+        $responseJson = $response->json();
+        $resultSet = new ResultSet($responseJson['results'], $responseJson['meta']);
+
+        $openActivity = OpenActivity::create($resultSet->results[0]);
 
         $this->assertInstanceOf('Ctct\Components\ResultSet', $resultSet);
-        $this->assertInstanceOf('Ctct\Components\Tracking\OpenActivity', $resultSet->results[0]);
+        $this->assertInstanceOf('Ctct\Components\Tracking\OpenActivity', $openActivity);
         $this->assertEquals("bGltaXQ9MyZuZXh0PTEzNTQ4MTcyMTA0MzA", $resultSet->next);
-        $this->assertEquals("EMAIL_OPEN", $resultSet->results[0]->activity_type);
-        $this->assertEquals(1100394165290, $resultSet->results[0]->campaign_id);
-        $this->assertEquals("86", $resultSet->results[0]->contact_id);
+        $this->assertEquals("EMAIL_OPEN", $openActivity->activity_type);
+        $this->assertEquals(1100394165290, $openActivity->campaign_id);
+        $this->assertEquals("86", $openActivity->contact_id);
         $this->assertEquals(
             "wizpie9e19a6d35ec249efa8fc3085721aa61d1354817227762990000@snoopy.roving.com",
-            $resultSet->results[0]->email_address
+            $openActivity->email_address
         );
 
-        $this->assertEquals("2012-12-06T13:07:11.839Z", $resultSet->results[0]->open_date);
+        $this->assertEquals("2012-12-06T13:07:11.839Z", $openActivity->open_date);
     }
 
     public function testGetSummary()
     {
-        $curlResponse = CurlResponse::create(JsonLoader::getSummary(), array('http_code' => 200));
-        $this->restClient->expects($this->once())
-            ->method('get')
-            ->with()
-            ->will($this->returnValue($curlResponse));
+        $response = self::$client->get('/');
 
-        $summary = $this->campaignTrackingService->getSummary('access_token', 1100394165290);
-
+        $summary = TrackingSummary::create($response->json());
         $this->assertInstanceOf('Ctct\Components\Tracking\TrackingSummary', $summary);
         $this->assertEquals(15, $summary->sends);
         $this->assertEquals(10, $summary->opens);
